# _Reversa_MasterSaaS/openapi/mastersaas-api.yaml
openapi: "3.1.0"
info:
  title: MasterSaaS API
  version: "1.0.0"
  description: |
    Especificação OpenAPI gerada pelo Reversa v1.2.14 a partir de engenharia reversa do frontend.
    O sistema atual opera em mock in-memory — esta spec define o contrato de backend a implementar.
    Autenticação delegada ao Supabase Auth; endpoints usam Bearer JWT emitido pelo Supabase.
servers:
  - url: https://api.mastersaas.com/v1
    description: Produção
  - url: http://localhost:3000/v1
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido pelo Supabase Auth (anon key ou service_role key)
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer

    # ─── Auth / Profiles ──────────────────────────────────────────────────────
...
    CommissionStatus:
      type: string
      enum: [pending, available, processing, paid, canceled, failed, refunded]
      description: |
        pending: aguardando holding 30 dias
        available: liberado para saque
        processing: aprovado pelo admin, pagamento em andamento
        paid: pago (terminal)
        canceled: cancelado (terminal)
        failed: falha no pagamento (pode retry)
        refunded: estornado (terminal)
    Commission:
      type: object
      required: [id, sale_id, affiliate_id, revenue, commission, rate_snapshot, sale_date, hold_until, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        sale_id:
          type: string
          format: uuid
        affiliate_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
          nullable: true
          description: "Snapshot da promoção ativa no momento da venda"
        revenue:
          type: number
        commission:
          type: number
          description: "revenue × rate_snapshot / 100"
        rate_snapshot:
          type: number
          description: "Taxa % no momento da venda — imutável"
        sale_date:
          type: string
          format: date-time
        hold_until:
          type: string
          format: date-time
          description: "sale_date + 30 dias"
        available_at:
          type: string
          format: date-time
          nullable: true
        paid_at:
          type: string
          format: date-time
          nullable: true
        canceled_at:
          type: string
          format: date-time
          nullable: true
        payment_id:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CommissionStatus"
        created_at:
          type: string
          format: date-time
    CommissionHistory:
      type: object
      required: [id, commission_id, from_status, to_status, changed_at]
      properties:
        id:
          type: string
          format: uuid
        commission_id:
          type: string
          format: uuid
        from_status:
          $ref: "#/components/schemas/CommissionStatus"
        to_status:
          $ref: "#/components/schemas/CommissionStatus"
        changed_by:
          type: string
          format: uuid
          nullable: true
        reason:
          type: string
          nullable: true
        changed_at:
          type: string
          format: date-time
    AffiliateSummary:
      type: object
      properties:
        balance_available:
          type: number
        balance_pending:
          type: number
        total_earned:
          type: number
        total_paid:
          type: number

    # ─── Finance / Withdrawals ────────────────────────────────────────────────
...
paths:
...
  /commissions:
    get:
      tags: [Commissions]
      summary: Lista comissões do afiliado autenticado
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CommissionStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Comissões do afiliado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Commission"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
                  summary:
                    $ref: "#/components/schemas/AffiliateSummary"
...
  /commissions/{id}/history:
    get:
      tags: [Commissions]
      summary: Trilha de auditoria de uma comissão
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Histórico de status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommissionHistory"

  # ── Withdrawals ───────────────────────────────────────────────────────────
...